+++ game/src/guicomponents/MissionCounter.h
@@ -11,12 +11,12 @@ public:
         if(_bounds.contains(cursorPosition)) {
             g->beginPath().setFillStyle(phantom::Colors::MIDNIGHTBLUE);
             for(unsigned i = 0; i < 5; ++i)
-                g->text(_bounds.origin.x, _bounds.origin.y + (i * (_bounds.size.y)), _bounds.size.y / 2, "fonts/arial.ttf", std::string("Blaat"));
+                g->text(_bounds.origin.x, _bounds.origin.y + (i * (_bounds.size.y)), static_cast<int>(_bounds.size.y / 2), "fonts/arial.ttf", std::string("Blaat"));
             g->stroke();
         }
         else {
             g->beginPath().setFillStyle(phantom::Colors::WHITE);
-            g->text(_bounds.origin.x, _bounds.origin.y, _bounds.size.y / 2, "fonts/arial.ttf", std::string("Missions: "));
+            g->text(_bounds.origin.x, _bounds.origin.y, static_cast<int>(_bounds.size.y / 2), "fonts/arial.ttf", std::string("Missions: "));
             g->stroke();
         }
     }
